@steveri 2023-11-28
> This looks really great, it will save tons of time debugging. I think since its a really big change, will require multiple repos to be updated, and the tapeout is this week we should hold off and merge it sometime next week if that is okay with you Steve?

Yes absolutely, that makes total sense to me, we will revisit this next week.