@steveri 2023-12-18
Okay so it looks as though this is working finally. Below are with-and-without comparisons of the daemon working with the aha-flow CI on the complete 3-hour "daily" regressions. Because of the daemon, dense-app "map" times are consistently lower, corresponding to the amount of time that would otherwise have been spent redundantly building the same 28x16 garnet circuit over and over...

#### WITHOUT DAEMON::
https://buildkite.com/stanford-aha/aha-flow/builds/9635
```
Fri Dec 15 16:52:06 2023
step                              total    compile     map       test
----------------------------  ---------  ---------  --------  ---------
garnet                             962
vec_elemadd_glb                     35         0         0        35
matmul_ikj_glb                      36         0         0        36
...
conv1_glb                          915       111       240       562
conv4_1_glb                       1406       449       688       268
conv4_x_glb                       1418       462       725       230
conv5_x_glb                       1224       430       680       112
```

#### WITH DAEMON
https://buildkite.com/stanford-aha/aha-flow/builds/9632
```
Fri Dec 15 12:18:12 2023
TIMING INFO
step                              total    compile      map       test
----------------------------  ---------  ---------  ---------  ---------
garnet                             940
vec_elemadd_glb                     38         0          0         38
matmul_ikj_glb                      36         0          0         36
...
conv1_glb                          992       134        159        698
conv4_1_glb                       1419       467        622        329
conv4_x_glb                       1396       486        634        275
conv5_x_glb                       1157       453        585        117
```