@steveri 2023-12-18
I will give reviewers a couple of days to mull this over, and then will merge the changes if no objections arise. Note that the changes are nondestructive in that `garnet.py` behavior will not change until/unless people use the new `--daemon` flag. Watch for separate upcoming aha-and garnet-repo pull-requests that will propose using `--daemon` for CI regressions