@steveri 2023-12-19
> Looks great Steve! I appreciate all the work you put in, it will save us so much time going forward

Thanks Jack. It's crazy how much work it took to do such a simple thing...